Fix bug in FUNcube telemetry parsing
For some weird reason, using this instead of lambda to check
if the frame type is correct didn't work.

Now we are also asking for __getitem__ (which is what we use)
instead of whether the frame is an instance of int. This seems more
robust.

(cherry picked from commit 739554562a06978eab3c8f86738cb27928ba0b32)