Update BER plots
It seems the FSK BER plot is much worse for low CN0 than last
time I generated it. I wonder with which parameters it was done
last time, but it seems I didn't commit that to git. Probably
it was made with an unrealistically low clock bandwidth.

In fact, I have checked that a plot quite similar to the old
one is obtained if a clock bandwidth of 0.02 instead of the
current default of 0.06 is used.