Fix bug in hdlc_framer
This fixes a bug in hdlc_framer that was introduced when porting
from Python 2 to Python 3. The buffer where the bits are appended
was a bytes in Python 3, which is inmutable. This has been changed
to a list.

This commit also adds a QA test of the hdlc_framer and hdlc_deframer
blocks.

This fixes #119