Add option to disable scrambling in CCSDS deframers
This is needed for BeliefSat-1, which uses CCSDS Reed-Solomon
frames but no scrambling. It is also good to have this in
general, as other satellites or systems might do the same.

The number of CCSDS framings for the SatYAML files has now
duplicated. This is a bit ugly, but for the time being will
remain like so.

This closes #90