Show version if gr_satellites --version is run
Due to the way that gr_satellites uses argparse, previously it
was necessary to run "gr_satellites SAT --version" (where SAT
is the name of a valid satellite) in order to show the version.
This was a bit confusing.

Now "gr_satellites --version" is also added as an special case
in order to achieve the expect result.

This closes #138