Add recording start time option in gr_satellites
This option makes the KISS file sink and the SatNOGS telemetry
submit blocks generate accurate timestamps for a recording by
counting the elapsed clock time from the beginning of execution
and adding that to the start time supplied as a parameter.

The recording needs to be played back at 1x speed for this option
to work correctly.