Remove vectorized implementations for encode_rs_8()
These depend on run-time detection of the CPU flags using the
cpu_features() function that was never copied over from libfec.
The bug didn't trigger on a build on x86_64, but it triggered
on x86 (see #167).

This bug was introduced in
5d49fea58481d5ec183c0fa390de0e7baff73942

To fix the bug, we simply remove the optimized implementations,
since this kind of implementation should be done in volk or
gnuradio.