Use datetime.isoformat() to parse recording timestamp
Instead of using datetime.strptime() with a "YYYY-MM-DD HH:MM:SS"
format, which needs quotes to prevent the space from splitting
the argument in two when used in the command line, we use
datetime.isoformat(). This supports using T as a separator instead
of a white space, and gives a more flexible syntax.

See #171