Add SatYAML parameters for CCSDS deframers
This eliminates all the framing variants of the CCSDS codes, leaving
only the CCSDS concatenated and the CCSDS Reed-Solomon framing. The
variants are specified by using the following parameters in the
SatYAML:

precoding: differential. If not specified, differential precoding is
  not used.
RS base: either "conventional" or "dual". This is mandatory
RS interleaving: interleave depth. This defaults to 1 if not specified.
scrambler: either "CCSDS" or "none". This defaults to "CCSDS" if not specified.
convolutional: one of "CCSDS", "NASA-DSN", "CCSDS uninverted", "NASA-DSN uninverted".
  This defaults to "CCSDS" if not specified.