Make --clk_limit parameter relative to samples per symbol
This parameter was always intended to be relative to the baudrate,
but it was absolute (or rather relative to the sample rate) due to
a misunderstanding of how the corresponding parameter of the
Symbol Sync block works. In fact, it was documented as relative
in the --help.

Following discussion in #196 I noticed that it was absolute and
decided to make it relative to the sample rate, maintaining corresponding
default value for 9k6 signals. The default value now has a different
effect for all the other baudrates.