Fix some build problems with lambdas in MSVC
One of the problems is caused by a ternary operator having
different lambdas in the "if" and "else" clauses. This gives some
problem regarding type conversion. The ternary operator has been
replaced by an if/else construct.

The other problem regards a constexpr variable that is not captured. The
constexpr variable has been set to static constexpr.

This closes #197