Add filter before quadrature demod in FSK demodulator
For IQ input, the FSK signal is quadrature demodulated. Currently,
the whole input bandwidth was fed into the quadrature demod. This
causes an unnecessarily large noise bandwidth to enter the non-linear
demod, giving suboptimal results.

A lowpass filter that cuts off according to Carson's rule has been
placed in front of the quadrature demod in order to improve the
demodulation.