Fix runtime errors caused by baudrate too high for sample rate
Some of the filters and the symbol sync block will give runtime
errors caused by parameters computed in cases when the samples
per symbol is too low.

This patch fixes the runtime errors by printing a warning and
resetting the baudrate to a non-problematic value. The demodulator
will not work, in any case, but it will not prevent the execution.

This closes #207