Initial port to GNU Radio 3.9
pybind11 bindings for all the C++ blocks have been generated and
other changes have been done as indicated in the OOT porting guide.
This now builds correctly with GNU Radio 3.9 and the tests pass.

There are still a few things to solve:
* The docstrings for the encode_rs and decode_rs bindings were
causing build fails and have been disabled
* Under pybind11, pmt_init_u8vector needs a List[Int] argument
when called from Python. There are many places in gr-satellites
where it is called with a bytes argument instead, so all these
places need to be update accordingly

See #215