Add GR_SATELLITES_SUBMIT_TLM env variable to control tlm submit
This reads the environment variable GR_SATELLITES_SUBMIT_TLM
from the gr-satellites core flowgraph. If the variable exists
and can be converted into an int, its truth value forces/disables
telemetry submission, ignoring the value in the config file.

This is used in the test.sh script to disable telemetry submission
as suggested in #214