Fix duplicated telemetry printing with --hexdump
For satellites with an "unknown" telemetry description, the
telemetry was printed twice if --hexdump was enabled. The cause
is that the hexdump sink that is used for "unknown" telemetry
was not considered a telemetry sink, and so it was not skipped
when running with --hexdump.

(cherry picked from commit 12e27d95a36594bb98429e841c3d570e7bb8003b)