Change default output path for file receiver to .
The previous default was /tmp, but in Windows there is no such
folder, so the thread running the file receiver would give an
exception and stop working unless the user gives a non-default
value.

Having '.' (the working directory) as default output path is
probably less problematic in a cross-platform environment.