Byte swap CSP header in some telemetry parsers
In ccc38e77961ebc3ca50d341b3b6b7d40734dc72c, the CSP header swap
in the AX100 and U482C deframers was removed (see #295 for the
rationale). As a consequence, the CSP header definition of the
telemetry parser was byte swapped.

However, the telemetry parsers of satellites that use CSP headers
but do not use the AX100 or U482C were not adjusted for this change.
This swaps (or removes the swap of) the CSP header in the parsers
for AAUSAT-4 (which should have a swap, and hadn't) and BY70-1
(which shouldn't have a swap but had).

The other satellite affected by this change is Quetzal-1, but its
parser does not need to be modified because it does not have a
swap and does not need one, so the parser was behaving incorrectly
before ccc38e77961ebc3ca50d341b3b6b7d40734dc72c.

(cherry picked from commit e7df8d018128a5b1424682a496a0398d282a87b6)