Add compatibility layer for PDU blocks
The pdu_to_tagged_stream and tagged_stream_to_pdu blocks have been
moved from gnuradio.blocks to gnuradio.pdu in GNU Radio 3.10. The
blocks are still available in the former location, but a deprecation
warning is printed if you use them from there.

This includes a compatibility layer, similar to grtypes.py, to import
the blocks from the appropriate location depending on the version of
GNU Radio.