Add CRC blocks
This adds C++ implementations of generic CRC blocks. The blocks
are CRC append, which computes the CRC of some PDU data and appends
it to the PDU, and CRC check, which checks if the CRC of a PDU is
correct.

(cherry picked from commit 476ccb6ad59045d730cdeb477267678fa289fb55)