Fix bug in kiss_server_sink
The port parameter of the socket_pdu block should be an str. Using an int
instead gives an error under GNU Radio 3.9 (and probably 3.10), since
pybind11 insists that the type must be correct. I think that under
GNU Radio 3.8 (with swig), an int parameter was converted silently to
an str, so we saw no problems.

This fixes #342