Add 8APSK Costas Loop block
This is an Costas Loop block for the 8APSK constellation used in
the QO-100 multimedia beacon.

This block uses control_loop from gr-digital, so now we need to
link the C++ library against gnuradio-digital, which requires finding
a bunch of gnuradio modules in CMake. Additionally, the libsndfile
headers are needed. The CI Docker images have been updated to add
libsndfile1-dev.