Add Fixed Length to PDU block
This is a C++ block that is intended to replaced the Python
fixedlen_tagger that is used in the Sync and create PDU family
of hierarchical blocks.

Rather than converting from a stream with tags to a tagged stream,
Fixed Length to PDU outputs PDUs directly, similarly to gr-pdu's
Tags to PDU. The main different with Tags to PDU is that the data
in the output PDUs may overlap.

The implementation of the block uses its own history circular buffer,
whose size is the packet length minus 1. Tags in the input are saved
to a linked list. When the packet corresponding to a tag ends in the
current input buffer, the packet is reconstructed potentially using
the data in the history circular buffer and the tag is removed from
the list.