Use .zip extension in QO-100 multimedia file receiver
Several of the file types transmitted by the QO-100 multimedia beacon
are zipped, but the filename doesn't include the .zip extension. Currently,
we were unzipping and overwriting the file upon completion. This gives
problems, because when the file is transmitted again, the chunks of the
zip file start overwriting the chunks of the unzipped file.

To avoid this, we append the .zip extension to those file types which
we know are zipped according to their file type. This makes the zipped
and unzipped files have different names.