qo100-multimedia-beacon: add support for bulletins
Bulletins are handled by a TCP server in the qo100_websockets_server.py
that listens in port 52002 and sends all the data it receives from a TCP
connection as a websocket message.

The file receiver uses the on_file() method to spin up a thread that tries
to connect to this TCP server to deliver the file contents. A thread is used
to prevent blocking if the TCP connection needs to wait (note that
qo100_websockets_server.py may not be running).

This also removes an extraneous #! in a comment in qo100_websockets_server.py.

(cherry picked from commit 19e58f951b2851ea865450c3ae0f8c629721f541)