Fix GEOSCAN-EDELVEIS 9k6 baudrate
The baudrate was set to 8k rather than 9k6 (which is what indicates
in the documentation) because that was what appeared in the PE0SAT's
recordings. However, these recordings didn't have the correct 48ksps
sample rate, so the baudrate measurement was wrong.

Recordings by DK3WN show that the baudrate is indeed 9k6.