Add support for Light-1 and BlueWalker 3
Light-1 and BlueWalker 3 both use the same coding (but the frame size
is different). This coding is identical to that used by Reaktor Hello
World except for the fact that a different 32bit syncword is used.

The reaktor_hello_world deframer now also serves as a Light-1 deframer
by using an optional parameter to specify the syncword. The Light-1
deframer has its own GRC block. In SatYAML, a new keywork Light-1 has
been added.