Try to fix qa_fixedlen_tagger flaky test
This test is failing sometimes because 5 rather than 6 PDUs have
been received at the end of the test. It might be that the cause
it that the final portion of the vector source data is not consumed
properly. This increases the length of the vector source data to
try to push all the important data through.