fsk_demodulator: invert symbols in non-IQ if deviation is negative
We use a negative deviation to indicate that the high tone encodes
the bit 0 and the low tone encodes the bit 1. With IQ input, this
is taken into account for quadrature demodulation. However, with
non-IQ input, which is already FM-demodulated, this wasn't taken
into account. In this case it is necessary to invert the stream
of symbols at some point.

See the discussion in
https://github.com/daniestevez/gr-satellites/discussions/307#discussioncomment-3872276

(cherry picked from commit 730ad51959d7c8a54e56147d6b127f280630ed34)