Add support for CAS-5A new image downlink format
This adds support for the new image downlink format that has been
recently enabled in CAS-5A. Compared to the old format, this adds
a sequence number, which makes it easier to control lost packets.

There doesn't seem to be a good way of discriminating between new
image packets and old image packets (new image packets start with
0x03, while old image packets can start with any value). Therefore,
the decoding of old image packets has been disabled in the SatYAML
file. The satellite is still capable of using both formats. The
SatYAML needs to be modified manually to process images in the old
format.