Fix bug in Doppler correction
This fixes a bug in the Doppler correction calculation. The subtraction
of nitems_written(0) and d_sample_t0 was carried out as unsigned, which
overflowed if d_sample_t0 > nitems_written(0), as would happen when the
time had been updated due to a tag not at the beginning of the current
buffer. This would cause the d_current_index to increase until the end
of the data and get stuck there.

Additionally, to account for cases in which a time update to "go to
the past" is received, a small routine has been added which moves the
d_current_index backwards as required.