Add default NORAD ID in Satellite Decoder GRC block
Otherwise a NORAD ID needs to be entered by the user even if
NORAD ID is not selected as the method to choosed the satellite.

Additionally, this sets "Satellite name" as the default method for
choosing the satellite, since it is more common than "YAML file".

(cherry picked from commit 91dbac5fe3e8a0b4e09d946839224ea8461ca21d)