Fix BER simulation
The BER simulation was broken because the order of the parameters
of the FFT block changed at some point, so fftshift was being applied
to the forward FFT (while no fftshift was the correct thing for
the BER calculation).

This also formats the BER simulation script to comply with pycodestyle.

The BER plots are updated after running the BER script. The FSK plot
has improved substantially due to some improvements done at some point
regarding low-pass filtering using Carson's rule before quadrature
demodulation.

This closes #434.