Solve sensitivity loss in AFSK demodulator in IQ mode
When running in IQ mode, the AFSK demodulator had the quadrature
demodulator directly connected to its input. A low-pass filter is
required, since quadrature demodulator is non-linear. This adds
a low-pass filter that filters to a bandwidth given by Carson's
bandwidth rule, which is the same method that the FSK demodulator
uses in IQ mode. An FM deviation of 3 kHz is assumed to compute
Carson's rule.

See #463.