Add BME WebSocket telmetry submitter
This adds support for sending telemetry to the BME WebSocket server,
which is used by MRC-100.

The implementation uses the Python websocket-client package (Debian
/Ubuntu package python3-websocket), which has been added as a dependency
in the installation instructions in the documentation and in the
debian package control file. The build-ubuntu CI action has been updated
to include python3-websocket in the Docker image.

The older BME HTTP telemetry server has been marked as deprecated,
since it is no longer operational.