Port Manchester Sync to C++
This ports the Manchester Sync block to C++, and changes the algorithm
that is used to detect the phase of the Manchester clock. The new
algorithm is based on maximum likelihood and works by blocks. The
metrics for a block of N samples are computed, and the output is formed
for that block is formed based on the comparison of those metrics.