CPU usage improvement for Selector block
The Selector block has a tricky implementation of the forecast()
function, because we want to be able to dump any items present on
inactive inputs, but forecast() does not have enough information
for us to decide if this is the case or to communication this need
to the runtime. This limitation causes that under some conditions
the block is called repeated with no items on any input, consuming
high CPU.

This adds handling for two special cases to limit the CPU usage.
If there is only one input, then we do not need to lie to the
runtime, because we do not need to dump items on other inputs. If
there are exactly two inputs and we are requested to produce one
sample, then we lie and say that we need exactly one sample in the
inactive input. This is not exactly right, because there might be
samples in the active input which we might process normally, however,
the hope is that if so, then there are multiple samples, in which
case a previous forecast() call requesting more output samples would
have "succeeded".

(cherry picked from commit e0657f2c5aae850f9c86a0f1fb52c3c74b1440ce)