Fix AF deviation in SALSAT yml file
See
https://github.com/daniestevez/gr-satellites/issues/463#issuecomment-1781439116

After measurement, it turns out that the AFSK tone deviation used
by SALSAT is 300 Hz instead of 600 Hz. In other words, the AF signal
is 1k2 MSK with a carrier of 1500 Hz. It is clear that the lower tone
is at 1200 Hz, since the packets start with a long lower tone.

The value of 600 Hz seems inherited from the SNET yml files, since
these satellites are similar to SALSAT. Maybe the value is also
wrong for them.

(cherry picked from commit 0160e77d0e474db3acd913660682b9064cb6e875)