Add --fm-deviation option to AFSK demodulator
This adds an --fm-deviation command line option to the AFSK demodulator.
It can be used in cases when the default of 3kHz is very different
from the true deviation (the deviation is only used to compute the
bandwidth for a lowpass filter using Carson's rule). An optional field
fm_deviation has been added to SatYAML. This field can be used to force
the FM deviation in case it is known. If the SatYAML file specifies
fm_deviation, then the --fm-deviation command line option has no effect.