Add HADES-D deframer
This adds a deframer for HADES-D. This satellite uses 50 baud FSK
with a deviation of approximately 550 Hz. Since the deviation is
much larger than the baudrate, the usual FSK demodulator in
gr-satellites won't work well for this signal. A custom demodulator
flowgraph based on the difference of the power detection of each FSK
tone will be added as an example.