Add time synchronization tag to Doppler correction block
This adds an option to synchronize the internal time of the
Doppler correction block to the current UTC time according to
the PC clock whenever the tag is received. It is intended to
be used with the packet_len tag in a burst transmitter for
uplink Doppler correction.

Also, this sets the filename argument to const, since it should
be a const std::string&.

(cherry picked from commit cd1d1216e00e32400aeb84b04880df9330bccb10)