Import pdu_set from gr.blocks and remove pdu_lambda
gr.pdu does not exist in GNU Radio 3.9. The blocks are in gr.block.
Moreover, pdu_lambda was added in GNU Radio 3.10, but it this case
it is trivial to remove it, since it is only acting as a passthrough.

(cherry picked from commit 621434e2d833bc20d0839405252311dee5fc530f)