Add packet length tag support to fixedlen_to_pdu
This adds variable packet length support to fixedlen_to_pdu (so now
the name of the block is almost a misnomer) by using an additional
and optional packet length tag. The length specified in fixedlen_to_pdu
is used both a maximum packet length (to size the internal history,
for instance), and also as default packet length if the packet length
tag is not present at the beginning of a packet or if this option is
not used.

The behaviour of the block remains the same if the packet length tag
is not specified (if it is the empty string), so this change is not
API breaking.

Also, the fixedlen_to_pdu implementation has been slightly refactored,
for instance by replacing the std::list used to hold the packet starts
by a std::vector. In order to efficiently remove elements from this
std::vector as they are consumed, the not consumed items are pushed back
to another empty vector and the two vectors are swapped at the end of
the loop.