Fix variable packet size bug in fixedlen_to_pdu
There was a bug in the way that variable packet size was implemented
in fixedlen_to_pdu. In the "packet starts after the write pointer"
case, a d_packetlen was replaced by info.length in the expression
to compute the position in the history where the packet data begins.
However, that d_packetlen in that expression is part of the history
size expression, so it does not stand for the packet length and it
should not be replaced by info.length.

This caused the wrong bytes to be output in some cases. The problem
happened more frequently with packet sizes that were close to the
maximum packet size, since in these cases it is easier to fall into
the case that had the bug.

This fixes the bug and replaces nitems_read(0) calls in multiple
expressions by a variable.