Add support for INHA-RoSAT
Decoder tested with
https://network.satnogs.org/observations/12818786/

This satellite is quite strange. An AX.25 G3RUH decoder produces
frames that are quite short but show the same structure and headers.
So it looks that the physical layer is AX.25, but the frame format
is not AX.25, since AX.25 headers are not present (some frames are
even too short to have AX.25 headers).